made variable un long int

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
 void print_binary(unsigned long int n)
 {
 	int digit;
-	int tmp;
+	unsigned long int tmp;
 	int printflag;
 
 	if (n == 0)
